fix: Moving falcon-server/graphlUtils to falcon-server-env package
