+++ packages/falcon-server-env/src/helpers/graphql.ts
@@ -1,4 +1,7 @@
 import {
+  DocumentNode,
+  Kind,
+  parse,
   GraphQLType,
   GraphQLNamedType,
   GraphQLResolveInfo,
@@ -9,11 +12,56 @@ import {
   ResponsePath
 } from 'graphql';
 
-export * from './schema';
-
 export const PATH_SEPARATOR: string = '.';
 export const TAG_SEPARATOR: string = ':';
 export const PARENT_KEYWORD: string = '$parent';
+
+export type RootFieldTypes = Record<string, string[]>;
+
+export interface RootTypeFieldsFn {
+  (typeDefs: string): RootFieldTypes;
+  (typeDefs: Array<string>): RootFieldTypes;
+}
+
+export const getRootTypeFields: RootTypeFieldsFn = (typeDefs: string | Array<string>): RootFieldTypes => {
+  const result: RootFieldTypes = {};
+  if (!typeDefs) {
+    return result;
+  }
+  const combinedTypeDefs: string = Array.isArray(typeDefs) ? typeDefs.join('\n') : typeDefs;
+
+  try {
+    const docNode: DocumentNode = parse(combinedTypeDefs);
+    docNode.definitions.forEach(definition => {
+      if (definition.kind !== Kind.OBJECT_TYPE_DEFINITION && definition.kind !== Kind.OBJECT_TYPE_EXTENSION) {
+        return;
+      }
+
+      const typeName: string = definition.name.value;
+      if (!['Query', 'Mutation'].includes(typeName)) {
+        return;
+      }
+
+      if (definition.fields) {
+        definition.fields.forEach(field => {
+          const fieldName = field.name.value;
+          if (!(typeName in result)) {
+            result[typeName] = [];
+          }
+
+          if (!result[typeName].includes(fieldName)) {
+            result[typeName].push(fieldName);
+          }
+        });
+      }
+    });
+  } catch (error) {
+    error.message = `Failed to get root type fields - ${error.message}`;
+    throw error;
+  }
+
+  return result;
+};
 
 /**
  * Generates a path-like string for the provided request
