+++ packages/falcon-server-env/src/helpers/index.ts
@@ -1,3 +1,4 @@
+export * from './graphql';
 export * from './html';
-export * from './url';
 export * from './proxyEndpoints';
+export * from './url';
