+++ packages/falcon-server/src/containers/ExtensionContainer.ts
@@ -10,13 +10,14 @@ import {
   GraphQLContext,
   GraphQLResolver,
   GraphQLResolverMap,
-  RemoteBackendConfig
+  RemoteBackendConfig,
+  getRootTypeFields,
+  RootFieldTypes
 } from '@deity/falcon-server-env';
 import { IResolvers } from 'apollo-server-koa';
 import { GraphQLResolveInfo } from 'graphql';
 import { mergeSchemas, makeExecutableSchema } from 'graphql-tools';
 import deepMerge from 'deepmerge';
-import { getRootTypeFields, RootFieldTypes } from '../graphqlUtils/schema';
 import { BackendConfig, ExtensionGraphQLConfig, ExtensionEntryMap } from '../types';
 import { BaseContainer } from './BaseContainer';
